refactor(peagen): rename protocols to transport
