+++ pkgs/standards/peagen/peagen/cli/commands/analysis.py
@@ -10,7 +10,7 @@ import typer
 from functools import partial
 
 from peagen.handlers.analysis_handler import analysis_handler
-from peagen.protocols import TASK_SUBMIT
+from peagen.transport import TASK_SUBMIT
 from peagen.transport.json_rpcschemas.task import SubmitParams, SubmitResult
 from peagen.cli.task_builder import _build_task as _generic_build_task
