+++ pkgs/standards/peagen/peagen/cli/commands/eval.py
@@ -21,7 +21,7 @@ from functools import partial
 from peagen._utils.config_loader import load_peagen_toml
 
 from peagen.handlers.eval_handler import eval_handler
-from peagen.protocols import TASK_SUBMIT
+from peagen.transport import TASK_SUBMIT
 from peagen.transport.json_rpcschemas.task import SubmitParams, SubmitResult
 from peagen.cli.task_builder import _build_task as _generic_build_task
