+++ pkgs/standards/peagen/peagen/cli/commands/evolve.py
@@ -14,7 +14,7 @@ from functools import partial
 from peagen.handlers.evolve_handler import evolve_handler
 from peagen.transport.jsonrpc_schemas import Status
 from peagen.core.validate_core import validate_evolve_spec
-from peagen.protocols import TASK_SUBMIT, TASK_GET
+from peagen.transport import TASK_SUBMIT, TASK_GET
 from peagen.transport.json_rpcschemas.task import (
     SubmitParams,
     SubmitResult,
