+++ pkgs/standards/peagen/peagen/cli/commands/keys.py
@@ -11,7 +11,7 @@ import typer
 
 from peagen.plugins.secret_drivers import AutoGpgDriver
 from peagen.core import keys_core
-from peagen.protocols import KEYS_UPLOAD, KEYS_DELETE, KEYS_FETCH
+from peagen.transport import KEYS_UPLOAD, KEYS_DELETE, KEYS_FETCH
 from peagen.transport.json_rpcschemas.keys import (
     UploadParams,
     DeleteParams,
