+++ pkgs/standards/peagen/peagen/cli/commands/mutate.py
@@ -13,7 +13,7 @@ import typer
 from functools import partial
 
 from peagen.handlers.mutate_handler import mutate_handler
-from peagen.protocols import TASK_SUBMIT
+from peagen.transport import TASK_SUBMIT
 from peagen.transport.json_rpcschemas.task import SubmitParams, SubmitResult
 from peagen.cli.task_builder import _build_task as _generic_build_task
