+++ pkgs/standards/peagen/peagen/cli/commands/templates.py
@@ -8,7 +8,7 @@ from peagen.cli.rpc_utils import rpc_post
 import typer
 
 from peagen.handlers.templates_handler import templates_handler
-from peagen.protocols import TASK_SUBMIT
+from peagen.transport import TASK_SUBMIT
 from peagen.transport.json_rpcschemas.task import SubmitResult
 from peagen.cli.task_builder import build_submit_params
