+++ pkgs/standards/peagen/peagen/cli/commands/validate.py
@@ -6,7 +6,7 @@ from typing import Any, Dict
 import typer
 
 from peagen.handlers.validate_handler import validate_handler
-from peagen.protocols import TASK_SUBMIT
+from peagen.transport import TASK_SUBMIT
 from peagen.transport.json_rpcschemas.task import SubmitResult
 from peagen.cli.task_builder import build_submit_params
 from peagen.cli.rpc_utils import rpc_post
