+++ pkgs/standards/peagen/peagen/gateway/rpc/tasks.py
@@ -4,9 +4,9 @@ import json
 import uuid
 import typing as t
 from peagen.transport.jsonrpc import RPCException
-from peagen.protocols.error_codes import Code as ErrorCode
+from peagen.transport.error_codes import Code as ErrorCode
 
-from peagen.protocols import (
+from peagen.transport import (
     TASK_SUBMIT,
     TASK_CANCEL,
     TASK_PAUSE,
