+++ pkgs/standards/peagen/peagen/transport/__init__.py
@@ -1,7 +1,7 @@
 """Typed JSON-RPC protocol primitives and method registry."""
 
 from peagen.transport.jsonrpc import RPCDispatcher
-from peagen.protocols import (
+from peagen.transport import (
     Request as RPCRequest,
     Response as RPCResponse,
     Error as RPCErrorData,
