+++ pkgs/standards/peagen/tests/unit/test_keys_cli.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 
 from peagen.cli.commands import keys as keys_mod
-from peagen.protocols import KEYS_UPLOAD, KEYS_DELETE, Response
+from peagen.transport import KEYS_UPLOAD, KEYS_DELETE, Response
 from peagen.transport.json_rpcschemas.keys import FetchResult
